Exit/requirements.txt
Flask==2.2.2
mysql-connector-python==9.3.0
